add cli package
