+++ distributions/nuxt-press/package.json
@@ -43,6 +43,7 @@
   },
   "dependencies": {
     "@nuxt/cli-edge": "latest",
+    "@nuxtpress/cli": "0.0.1-beta.4",
     "@nuxtpress/docs": "0.0.1-beta.4",
     "consola": "^2.10.1"
   },
