+++ packages/cli/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@nuxtpress/cli",
+  "version": "0.0.1-beta.4",
+  "main": "src/index.js",
+  "repository": "https://github.com/nuxt/press",
+  "license": "MIT",
+  "files": [
+    "src"
+  ]
+}
